// src/shared/api/generated/schemas/workspacesSingleWorkspaceResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Smetchik Pro API
 * HTTP API сервиса smethick-pro.
 * OpenAPI spec version: 1.0
 */
import type { WorkspacesWorkspaceResponse } from './workspacesWorkspaceResponse';

export interface WorkspacesSingleWorkspaceResponse {
  workspace?: WorkspacesWorkspaceResponse;
}